Add 'alias.rs', refactor 'shell.rs'
